frontEnd/requirements.txt
streamlit==1.20.0
requests==2.28.2
python-dotenv==1.0.0
extra-streamlit-components==0.1.56
streamlit-extras==0.2.6
pandas==1.5.3
altair==4.2.2